Adding DEBUG mode for saving database locally

+++ tests/conftest.py
@@ -4,19 +4,26 @@ import os
 import tempfile
 import logging
 
+from pathlib import Path
+
 import pytest
 from project_tracking import create_app, api
 from project_tracking.database import get_session, init_db, close_db
 
-
+logger = logging.getLogger(__name__)
 
 @pytest.fixture
 def app():
-    db_fd, db_path = tempfile.mkstemp()
+    if os.getenv('DEBUG'):
+        db_path = Path(os.path.join("instance", "test_db.sql"))
+        db_path.unlink(missing_ok=True)
+        logger.debug("DB is here %s", db_path)
+    else:
+        db_fd, db_path = tempfile.mkstemp()
 
     app = create_app({
         'TESTING': True,
-        'SQLALCHEMY_DATABASE_URI': "sqlite:///{}".format(db_path),
+        'SQLALCHEMY_DATABASE_URI': f"sqlite:///{db_path}",
     })
 
     with app.app_context():
@@ -24,24 +31,29 @@ def app():
 
     yield app
 
-    os.close(db_fd)
-    os.unlink(db_path)
+    if not os.getenv('DEBUG'):
+        os.unlink(db_path)
+        os.close(db_fd)
 
 @pytest.fixture
 def not_app_db():
-    db_fd, db_path = tempfile.mkstemp()
+    if os.getenv('DEBUG'):
+        db_path = Path(os.path.join("instance", "test_db.sql"))
+        db_path.unlink(missing_ok=True)
+        logger.debug("DB is here %s", db_path)
+    else:
+        db_fd, db_path = tempfile.mkstemp()
 
-    db = get_session(no_app=True, db_uri="sqlite:///{}".format(db_path))
+    db = get_session(no_app=True, db_uri=f'sqlite:///{db_path}')
     init_db()
-
-    print(db_path)
 
     try:
         yield db
     finally:
         close_db(no_app=True)
-        os.close(db_fd)
-        # os.unlink(db_path)
+        if not os.getenv('DEBUG'):
+            os.unlink(db_path)
+            os.close(db_fd)
 
 
 @pytest.fixture
